Rename module ring -> ring_algorithms
